+++ passive/s3.js
@@ -1,21 +1,28 @@
 // S3 bucket finder by alishasinghania09@gmail.com
 
-function scan(ps, msg, src) {
-  // populate some parameters which will be needed if s3 bucket url is present
-  var alertRisk = 1;
-  var alertConfidence = 3;
-  var alertTitle = "s3 Bucket URL";
-  var alertDesc = "s3 Bucket URL found in response.";
-  var alertSolution =
-    "Remove s3 Buckets name from response or make sure the permissions in bucket are configured properly.";
-  var cweId = 200;
-  var wascId = 13;
+var ScanRuleMetadata = Java.type(
+  "org.zaproxy.addon.commonlib.scanrules.ScanRuleMetadata"
+);
 
+function getMetadata() {
+  return ScanRuleMetadata.fromYaml(`
+id: 100036
+name: Information Disclosure - Amazon S3 Bucket URL
+description: An Amazon S3 bucket URL was found in the HTTP response body.
+solution: Remove S3 Bucket names from the response or ensure that the permissions in bucket are configured properly.
+risk: low
+confidence: high
+cweId: 200  # CWE-200: Exposure of Sensitive Information to an Unauthorized Actor
+wascId: 13  # WASC-13: Information Leakage
+status: alpha
+codeLink: https://github.com/zaproxy/community-scripts/blob/main/passive/s3.js
+helpLink: https://www.zaproxy.org/docs/desktop/addons/community-scripts/
+`);
+}
+
+function scan(helper, msg, src) {
   // the regex for s3 bucket url and it must appear within /( and )/g
   var re = /((s3:\\[a-zA-Z0-9-\.\\_]+)|((s3-|s3\.)?(.*)\.amazonaws\.com))/g;
-
-  // we need to set the url variable to the request or we cant track the alert later
-  var url = msg.getRequestHeader().getURI().toString();
 
   // If the file type is image jpeg/png , then the scan will be skipped
   var contenttype = msg.getResponseHeader().getHeader("Content-Type");
@@ -39,21 +46,12 @@ function scan(ps, msg, src) {
         founds3bucket.push(buckets[0]);
       }
       //raise the alert
-      ps.raiseAlert(
-        alertRisk,
-        alertConfidence,
-        alertTitle,
-        alertDesc,
-        url,
-        "",
-        "",
-        founds3bucket.toString(),
-        alertSolution,
-        "",
-        cweId,
-        wascId,
-        msg
-      );
+      helper
+        .newAlert()
+        .setEvidence(founds3bucket[0])
+        .setOtherInfo(`Other instances: ${founds3bucket.slice(1).toString()}`)
+        .setMessage(msg)
+        .raise();
     }
   }
 }